move error to a module
