+++ Cargo.toml
@@ -11,3 +11,4 @@ serde_json = "1.0.107"
 tokio = { version = "1.2", features = ["full"] }
 warp = "0.3"
 log = "0.4.20"
+handle-errors = { path = "handle-errors" }
