+++ src/main.rs
@@ -1,173 +1,15 @@
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::sync::Arc;
-use tokio::sync::RwLock;
-use warp::{
-    http::Method,
-    http::StatusCode,
-    Filter,
-};
+#![warn(clippy::all)]
 
-#[derive(Debug, Clone)]
-struct Store {
-    questions: Arc<RwLock<HashMap<QuestionId, Question>>>,
-    answers: Arc<RwLock<HashMap<AnswerId, Answer>>>,
-}
+use handle_errors::return_error;
+use warp::{http::Method, Filter};
 
-impl Store {
-    fn new() -> Self {
-        Store {
-            questions: Arc::new(RwLock::new(Self::init())),
-            answers: Arc::new(RwLock::new(HashMap::new())),
-        }
-    }
-    fn init() -> HashMap<QuestionId, Question> {
-        let file = include_str!("../questions.json");
-        serde_json::from_str(file).expect("can't read questions.json")
-    }
-}
-
-#[derive(Deserialize, Debug, Serialize, Clone)]
-struct Question {
-    id: QuestionId,
-    title: String,
-    content: String,
-    tags: Option<Vec<String>>,
-}
-
-#[derive(Deserialize, Debug, Serialize, Clone, Eq, PartialEq, Hash)]
-struct QuestionId(String);
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-struct Answer {
-    id: AnswerId,
-    content: String,
-    question_id: QuestionId,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
-struct AnswerId(String);
-
-#[derive(Debug)]
-struct Pagination {
-    start: usize,
-    end: usize,
-}
-
-async fn get_questions(
-    params: HashMap<String, String>,
-    store: Store,
-) -> Result<impl warp::Reply, warp::Rejection> {
-    if !params.is_empty() {
-        let pagination = extract_pagination(params)?;
-        let res: Vec<Question> = store.questions.read().await.values().cloned().collect();
-        let (start, end) = prepare_pagination(pagination.start, pagination.end, res.len());
-        let res = &res[start..end];
-        Ok(warp::reply::json(&res))
-    } else {
-        let res: Vec<Question> = store.questions.read().await.values().cloned().collect();
-        Ok(warp::reply::json(&res))
-    }
-}
-
-async fn add_question(
-    store: Store,
-    question: Question,
-) -> Result<impl warp::Reply, warp::Rejection> {
-    store
-        .questions
-        .write()
-        .await
-        .insert(question.id.clone(), question);
-
-    Ok(warp::reply::with_status("Question added", StatusCode::OK))
-}
-
-async fn update_question(
-    id: String,
-    store: Store,
-    question: Question,
-) -> Result<impl warp::Reply, warp::Rejection> {
-    match store.questions.write().await.get_mut(&QuestionId(id)) {
-        Some(q) => *q = question,
-        None => return Err(warp::reject::custom(error::Error::QuestionNotFound)),
-    }
-
-    Ok(warp::reply::with_status("Question updated", StatusCode::OK))
-}
-
-async fn delete_question(
-    id: String,
-    store: Store,
-) -> Result<impl warp::Reply, warp::Rejection> {
-    match store.questions.write().await.remove(&QuestionId(id)) {
-        Some(_) => {
-            Ok(
-                warp::reply::with_status(
-                    "Question deleted",
-                    StatusCode::OK,
-                )
-            )
-        },
-        None => return Err(warp::reject::custom(error::Error::QuestionNotFound)),
-    }
-}
-
-async fn add_answer(
-    store: Store,
-    params: HashMap<String, String>,
-) -> Result<impl warp::Reply, warp::Rejection> {
-    let answer = Answer {
-        id: AnswerId("1".to_string()),
-        content: params.get("content").unwrap().to_string(),
-        question_id: QuestionId(
-            params.get("questionId").unwrap().to_string()
-        ),
-    };
-
-    store.answers.write().await.insert(answer.id.clone(), answer);
-
-    Ok(warp::reply::with_status("Answer added", StatusCode::OK))
-}
-
-fn prepare_pagination(start: usize, end: usize, length: usize) -> (usize, usize) {
-    if length == 1 {
-        (0, 1)
-    } else if start > end {
-        if end < length {
-            (0, end)
-        } else {
-            (0, length)
-        }
-    } else if end > length {
-        (start, length)
-    } else {
-        (start, end)
-    }
-}
-
-fn extract_pagination(params: HashMap<String, String>) -> Result<Pagination, error::Error> {
-    if params.contains_key("start") && params.contains_key("end") {
-        return Ok(Pagination {
-            start: params
-                .get("start")
-                .unwrap()
-                .parse::<usize>()
-                .map_err(error::Error::ParseError)?,
-            end: params
-                .get("end")
-                .unwrap()
-                .parse::<usize>()
-                .map_err(error::Error::ParseError)?,
-        });
-    }
-
-    Err(error::Error::MissingParameters)
-}
+mod store;
+mod types;
+mod routes;
 
 #[tokio::main]
 async fn main() {
-    let store = Store::new();
+    let store = store::Store::new();
     let store_filter = warp::any().map(move || store.clone());
 
     let cors = warp::cors()
@@ -180,14 +22,14 @@ async fn main() {
         .and(warp::path::end())
         .and(warp::query())
         .and(store_filter.clone())
-        .and_then(get_questions);
+        .and_then(routes::question::get_questions);
 
     let add_question = warp::post()
         .and(warp::path("questions"))
         .and(warp::path::end())
         .and(store_filter.clone())
         .and(warp::body::json())
-        .and_then(add_question);
+        .and_then(routes::question::add_question);
 
     let update_question = warp::put()
         .and(warp::path("questions"))
@@ -195,21 +37,21 @@ async fn main() {
         .and(warp::path::end())
         .and(store_filter.clone())
         .and(warp::body::json())
-        .and_then(update_question);
+        .and_then(routes::question::update_question);
 
     let delete_question = warp::delete()
         .and(warp::path("questions"))
         .and(warp::path::param::<String>())
         .and(warp::path::end())
         .and(store_filter.clone())
-        .and_then(delete_question);
+        .and_then(routes::question::delete_question);
 
     let add_answer = warp::post()
         .and(warp::path("answers"))
         .and(warp::path::end())
         .and(store_filter.clone())
         .and(warp::body::form())
-        .and_then(add_answer);
+        .and_then(routes::answer::add_answer);
 
     let routes = get_questions
         .or(add_question)
@@ -217,66 +59,7 @@ async fn main() {
         .or(update_question)
         .or(delete_question)
         .with(cors)
-        .recover(error::return_error);
+        .recover(return_error);
 
     warp::serve(routes).run(([127, 0, 0, 1], 3030)).await;
-}
-
-mod error {
-    use warp::{
-        filters::{
-            body::BodyDeserializeError,
-            cors::CorsForbidden,
-        },
-        reject::Reject,
-        Rejection,
-        Reply,
-        http::StatusCode
-    };
-    use std::fmt::Formatter;
-
-    #[derive(Debug)]
-    pub enum Error {
-        ParseError(std::num::ParseIntError),
-        MissingParameters,
-        QuestionNotFound,
-    }
-
-    impl std::fmt::Display for Error {
-        fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-            match *self {
-                Error::ParseError(ref err) => {
-                    write!(f, "Cannot parse paramter: {}", err)
-                }
-                Error::MissingParameters => write!(f, "Missing parameter"),
-                Error::QuestionNotFound => write!(f, "Question not found"),
-            }
-        }
-    }
-
-    impl Reject for Error {}
-
-    pub async fn return_error(r: Rejection) -> Result<impl Reply, Rejection> {
-        if let Some(error) = r.find::<Error>() {
-            Ok(warp::reply::with_status(
-                error.to_string(),
-                StatusCode::RANGE_NOT_SATISFIABLE,
-            ))
-        } else if let Some(error) = r.find::<CorsForbidden>() {
-            Ok(warp::reply::with_status(
-                error.to_string(),
-                StatusCode::FORBIDDEN,
-            ))
-        } else if let Some(error) = r.find::<BodyDeserializeError>() {
-            Ok(warp::reply::with_status(
-                error.to_string(),
-                StatusCode::UNPROCESSABLE_ENTITY,
-            ))
-        } else {
-            Ok(warp::reply::with_status(
-                "Route not found".to_string(),
-                StatusCode::NOT_FOUND,
-            ))
-        }
-    }
 }
