+++ src/types/mod.rs
@@ -1,0 +1,3 @@
+pub mod question;
+pub mod answer;
+pub mod pagination;
